move add post modal into renderbutton component
