+++ src/components/Dashboard/DashboardPost/AddPost.tsx
@@ -13,25 +13,26 @@ const handleSubElementClick = (e: MouseEvent) => {
 };
 
 const AddPost: FC<AddPostProp> = ({ isOpen, onClose, addPostRoot }) => {
+    if (!isOpen || !addPostRoot) return null
+
     return (
-        <>
-            {isOpen && addPostRoot && (
+        <>(
+            <div
+                className="fixed flex justify-center items-center w-screen inset-0 bg-gray-800 bg-opacity-50"
+                onClick={onClose}
+            >
                 <div
-                    className="fixed flex justify-center items-center w-screen inset-0 bg-gray-800 bg-opacity-50"
-                    onClick={onClose}
+                    onClick={handleSubElementClick}
+                    className={`max-w-lg bg-white lg:m-6 rounded-lg lg:p-6 z-20 transform ${isOpen ? "translate-x-0" : "-translate-x-full"
+                        } transition-transform duration-1000 ease-linear`}
                 >
-                    <div
-                        onClick={handleSubElementClick}
-                        className={`max-w-lg bg-white lg:m-6 rounded-lg lg:p-6 z-20 transform ${isOpen ? "translate-x-0" : "-translate-x-full"
-                            } transition-transform duration-1000 ease-linear`}
-                    >
-                        <h1 className="font-bold text-center uppercase text-2xl">
-                            Add new Post
-                        </h1>
-                        <AddPostForm onClose={onClose} />
-                    </div>
+                    <h1 className="font-bold text-center uppercase text-2xl">
+                        Add new Post
+                    </h1>
+                    <AddPostForm onClose={onClose} />
                 </div>
-            )}
+            </div>
+            )
         </>
     );
 };
