+++ src/components/Dashboard/DashboardTable/RenderButtons.tsx
@@ -1,46 +1,65 @@
-import { FC } from "react";
+'use client'
+import { FC, useRef, useState } from "react";
+import AddPost from "../DashboardPost/AddPost";
 type renderButtonProps = {
     tab: string;
-    toggleAddPost: () => void;
 }
 
-const renderAddButton: FC<renderButtonProps> = ({ tab, toggleAddPost }) => {
-    if (tab === "Users") {
-        return (
-            <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
-                Add User
-            </button>
-        );
-    } else if (tab === "Posts") {
-        return (
-            <button
-                className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl"
-                onClick={toggleAddPost}
-            >
-                Add post
-            </button>
-        );
-    } else if (tab === "Events") {
-        return (
-            <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
-                Add event
-            </button>
-        );
-    } else if (tab === "Comments") {
-        return (
-            <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
-                Add Comment
-            </button>
-        );
-    } else if (tab === "Categories") {
-        return (
-            <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
-                Add Category
-            </button>
-        );
-    }
+const RenderAddButton: FC<renderButtonProps> = ({ tab }) => {
+    const [isPostOpen, setPostOpen] = useState(false);
+    const postOpenRef = useRef<HTMLDivElement | null>(null);
 
-    return null; // Return null if the tab doesn't match any conditions
+    const toggleAddPost = () => {
+        setPostOpen((prevState) => !prevState);
+    };
+
+    return (
+        <div>
+            <div ref={postOpenRef} id="addPost-root" />
+            <AddPost
+                isOpen={isPostOpen}
+                onClose={toggleAddPost}
+                addPostRoot={postOpenRef.current}
+            />
+            {
+                (tab === "Users") && (<button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
+                    Add User
+                </button>)
+            }
+            {
+                (tab === "Posts") && <button
+                    className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl"
+                    onClick={toggleAddPost}
+                >
+                    Add post
+                </button>
+            }
+            {
+                (tab === "Event") && (
+                    <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
+                        Add event
+                    </button>
+                )
+            }
+
+            {
+                (tab === "Comments") && (
+                    <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
+                        Add Comment
+                    </button>
+                )
+            }
+            {
+                (tab === "Categories") && (
+                    <button className="bg-colorPrimary px-4 py-2 text-white hover:bg-green-700 rounded-xl">
+                        Add Category
+                    </button>
+                )
+            }
+
+        </div>
+    )
+
 };
 
-export default renderAddButton
+export default RenderAddButton
